Separated server synchronization into it's own Service
This commit abstracts server sync to it's own service. We have an interface MyPaceServerSyncService which has the blueprint of what server sync implementations should implement. We then implement MyPlaceFirebaseSyncService which deals with pushing MyPlace create events to firebase. MyPlaceServerSyncImpl is the one that chooses which will be the main implementation the app uses.For example, if we have implementation of MyPlaceFirebaseSyncService and MyPlaceMongoDBSyncService and MyPlaceMySqlSyncService with different backend, MyPlaceServerSyncImpl takes an argument of MyPlaceSyncService, and we can choose one of those options as the argument for it, for example MyPlaceServerSyncImpl(MyPlaceFirebaseSyncService) will have the app using firebase as the backend