Extracted more functions refactored all methods in MyPlacesDao to return Observable
This commit does a whole lot of things 😅 yikes. But mostly, it revolves making the methods from MyPlacesDao all return observables. The methods that didn't have return types in this case are update, insert and delete. Their returning of Observables means that it's possible to react to their events being completed successfully or failing. I also refactored the code that used MyPlacesDao, which is scattered all other. Functionality remains the same,the code is a bit easier to understand and friendlier to test